refactor(frontend): split SignalsPanel into per-component files (#61 eje 2)
Split the 1330-line SignalsPanel.jsx monolith into focused per-component
files under frontend/src/components/signals/:

- helpers.js — PAIRS, INTERVALS, fmt* helpers
- ConfigBadge.jsx — ConfigBadge + StatusBadge
- ToggleSwitch.jsx
- CapitalConfig.jsx
- ConfigForm.jsx
- ConfigsList.jsx
- SignalsList.jsx
- SimTradesList.jsx (incl. SimTradeReview + StopMovesDetail)
- RealTradesSection.jsx
- ComparisonView.jsx

SignalsPanel.jsx is now a thin orchestrator (1330 → 210 LOC) that
imports the sub-components and owns the data-fetching/state. No
behavioural or visual changes.

This unblocks adding tooltips, table density tweaks, and skeleton
states without touching a 1300-line file.
