+++ frontend/src/components/SignalsPanel.jsx
@@ -1,1128 +1,12 @@
-import React, { useState, useEffect, useCallback, useMemo } from 'react'
+import { useState, useEffect, useCallback } from 'react'
 import { apiFetch } from '../auth/apiFetch'
-import TradeReviewChart from './TradeReviewChart'
+import { ConfigForm } from './signals/ConfigForm'
+import { ConfigsList } from './signals/ConfigsList'
+import { SignalsList } from './signals/SignalsList'
+import { SimTradesList } from './signals/SimTradesList'
+import { RealTradesSection } from './signals/RealTradesSection'
+import { ComparisonView } from './signals/ComparisonView'
 
-const PAIRS = [
-  'BTCUSDT', 'ETHUSDT', 'BNBUSDT', 'SOLUSDT', 'XRPUSDT',
-  'ADAUSDT', 'DOGEUSDT', 'AVAXUSDT', 'DOTUSDT', 'MATICUSDT',
-]
-
-const INTERVALS = [
-  { value: '1h', label: '1 Hour' },
-  { value: '4h', label: '4 Hours' },
-  { value: '1d', label: '1 Day' },
-  { value: '1w', label: '1 Week' },
-  { value: '1M', label: '1 Month' },
-]
-
-function fmtNum(v, digits = 2) {
-  if (v === null || v === undefined) return '—'
-  return Number(v).toFixed(digits)
-}
-
-function fmtMoney(v) {
-  if (v === null || v === undefined) return '—'
-  return new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD', maximumFractionDigits: 2 }).format(Number(v))
-}
-
-function fmtTime(ms) {
-  if (!ms) return '—'
-  return new Date(Number(ms)).toLocaleString()
-}
-
-function fmtIso(iso) {
-  if (!iso) return '—'
-  return new Date(iso).toLocaleString()
-}
-
-function fmtConfigParams(raw) {
-  if (!raw) return ''
-  try {
-    const obj = typeof raw === 'string' ? JSON.parse(raw) : raw
-    return Object.entries(obj).map(([k, v]) => `${k}: ${v}`).join('\n')
-  } catch { return '' }
-}
-
-function ConfigBadge({ configId, strategy, params }) {
-  const lines = fmtConfigParams(params)
-  return (
-    <span style={{ position: 'relative', display: 'inline-block' }} className="config-badge-wrap">
-      <span
-        style={{
-          display: 'inline-flex', alignItems: 'center', gap: 4,
-          padding: '2px 8px', borderRadius: 'var(--radius-sm)',
-          background: 'var(--bg-elevated)', border: '1px solid var(--border-default)',
-          fontSize: '0.75rem', cursor: lines ? 'default' : 'default',
-          whiteSpace: 'nowrap',
-        }}
-      >
-        <span style={{ fontFamily: 'var(--font-mono)', color: 'var(--text-muted)', fontWeight: 600 }}>#{configId}</span>
-        <span style={{ color: 'var(--text-secondary)' }}>{strategy}</span>
-        {lines && <span style={{ color: 'var(--text-muted)', fontSize: '0.65rem' }}>ⓘ</span>}
-      </span>
-      {lines && (
-        <span className="config-badge-popover">
-          {lines.split('\n').map((l, i) => <span key={i} style={{ display: 'block' }}>{l}</span>)}
-        </span>
-      )}
-    </span>
-  )
-}
-
-/* ---- Status badge ---- */
-function StatusBadge({ status }) {
-  const colors = {
-    open: 'var(--color-success)',
-    pending_entry: 'var(--color-warning)',
-    closed: 'var(--text-muted)',
-    active: 'var(--color-success)',
-    pending: 'var(--color-warning)',
-  }
-  const color = colors[status] || 'var(--text-secondary)'
-  return (
-    <span style={{
-      display: 'inline-block', padding: '2px 8px', borderRadius: 'var(--radius-sm)',
-      background: `${color}22`, color, fontSize: '0.75rem', fontWeight: 600,
-      textTransform: 'uppercase', letterSpacing: '0.03em',
-    }}>
-      {status?.replace('_', ' ') || '—'}
-    </span>
-  )
-}
-
-/* ---- Toggle switch ---- */
-function ToggleSwitch({ checked, onChange, disabled, ariaLabel }) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={!!checked}
-      aria-label={ariaLabel}
-      aria-disabled={disabled || undefined}
-      disabled={disabled}
-      onClick={() => !disabled && onChange(!checked)}
-      style={{
-        width: 44, height: 24, borderRadius: 12, border: 'none', cursor: disabled ? 'default' : 'pointer',
-        background: checked ? 'var(--color-success)' : 'var(--border-strong)',
-        position: 'relative', transition: 'background var(--transition-fast)',
-        opacity: disabled ? 0.5 : 1,
-      }}
-    >
-      <span style={{
-        position: 'absolute', top: 2, left: checked ? 22 : 2,
-        width: 20, height: 20, borderRadius: '50%', background: 'var(--text-primary)',
-        transition: 'left var(--transition-fast)',
-      }} />
-    </button>
-  )
-}
-
-/* ---- Capital mode selector ---- */
-function CapitalConfig({ portfolio, setPortfolio, leverage, setLeverage, investedAmount, setInvestedAmount, mode, setMode, disabled }) {
-  return (
-    <div>
-      <div style={{ display: 'flex', gap: 'var(--space-2)', marginBottom: 'var(--space-3)' }}>
-        <button
-          type="button"
-          className={`btn btn-sm ${mode === 'leverage' ? 'btn-primary' : 'btn-secondary'}`}
-          onClick={() => setMode('leverage')}
-          disabled={disabled}
-        >Portfolio + Leverage</button>
-        <button
-          type="button"
-          className={`btn btn-sm ${mode === 'invested' ? 'btn-primary' : 'btn-secondary'}`}
-          onClick={() => setMode('invested')}
-          disabled={disabled}
-        >Portfolio + Invested</button>
-      </div>
-      <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 'var(--space-3)' }}>
-        <div className="form-group">
-          <label className="form-label">Portfolio (USD)</label>
-          <input type="number" className="form-control" value={portfolio} min={1} step={100}
-            onChange={e => setPortfolio(parseFloat(e.target.value) || 0)} disabled={disabled} />
-        </div>
-        {mode === 'leverage' ? (
-          <div className="form-group">
-            <label className="form-label">Leverage</label>
-            <input type="number" className="form-control" value={leverage} min={0.1} step={0.1}
-              onChange={e => setLeverage(parseFloat(e.target.value) || 1)} disabled={disabled} />
-          </div>
-        ) : (
-          <div className="form-group">
-            <label className="form-label">Invested Amount (USD)</label>
-            <input type="number" className="form-control" value={investedAmount} min={1} step={100}
-              onChange={e => setInvestedAmount(parseFloat(e.target.value) || 0)} disabled={disabled} />
-          </div>
-        )}
-      </div>
-    </div>
-  )
-}
-
-/* ---- Config creation form ---- */
-function ConfigForm({ strategies, onCreated }) {
-  const [symbol, setSymbol] = useState('BTCUSDT')
-  const [interval, setInterval] = useState('1d')
-  const [selectedStrat, setSelectedStrat] = useState('')
-  const [paramValues, setParamValues] = useState({})
-  const [costBps, setCostBps] = useState(10)
-  const [maintenanceMarginPct, setMaintenanceMarginPct] = useState(0.005)
-  const [portfolio, setPortfolio] = useState(10000)
-  const [leverage, setLeverage] = useState(1)
-  const [investedAmount, setInvestedAmount] = useState(10000)
-  const [capitalMode, setCapitalMode] = useState('leverage')
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState(null)
-
-  useEffect(() => {
-    if (strategies.length > 0 && !selectedStrat) {
-      const first = strategies[0]
-      setSelectedStrat(first.name)
-      const defaults = {}
-      for (const p of first.parameters ?? []) defaults[p.name] = p.default
-      setParamValues(defaults)
-    }
-  }, [strategies, selectedStrat])
-
-  const handleStrategyChange = e => {
-    const name = e.target.value
-    setSelectedStrat(name)
-    const strat = strategies.find(s => s.name === name)
-    if (strat) {
-      const defaults = {}
-      for (const p of strat.parameters ?? []) defaults[p.name] = p.default
-      setParamValues(defaults)
-    }
-  }
-
-  const currentStrat = strategies.find(s => s.name === selectedStrat)
-
-  const handleCreate = async () => {
-    setLoading(true)
-    setError(null)
-    try {
-      const body = {
-        symbol, interval, strategy: selectedStrat,
-        params: paramValues,
-        cost_bps: costBps,
-        maintenance_margin_pct: maintenanceMarginPct,
-        initial_portfolio: portfolio,
-        ...(capitalMode === 'leverage'
-          ? { leverage, invested_amount: null }
-          : { invested_amount: investedAmount, leverage: null }),
-      }
-      const res = await apiFetch('/api/signals/configs', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      })
-      if (!res.ok) {
-        const err = await res.json().catch(() => ({}))
-        setError(err.detail ?? `HTTP ${res.status}`)
-        return
-      }
-      onCreated()
-    } catch (e) {
-      setError(e.message)
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--space-4)' }}>
-      <div style={{ display: 'grid', gridTemplateColumns: 'repeat(3, 1fr)', gap: 'var(--space-3)' }}>
-        <div className="form-group">
-          <label className="form-label">Pair</label>
-          <select className="form-control" value={symbol} onChange={e => setSymbol(e.target.value)} disabled={loading}>
-            {PAIRS.map(p => <option key={p} value={p}>{p}</option>)}
-          </select>
-        </div>
-        <div className="form-group">
-          <label className="form-label">Interval</label>
-          <select className="form-control" value={interval} onChange={e => setInterval(e.target.value)} disabled={loading}>
-            {INTERVALS.map(iv => <option key={iv.value} value={iv.value}>{iv.label}</option>)}
-          </select>
-        </div>
-        <div className="form-group">
-          <label className="form-label">Strategy</label>
-          {strategies.length === 0 ? (
-            <div style={{ padding: '8px 12px', background: 'var(--bg-input)', border: '1px solid var(--border-default)', borderRadius: 'var(--radius-sm)', color: 'var(--text-muted)', fontSize: '0.83rem' }}>
-              Loading strategies…
-            </div>
-          ) : (
-            <select className="form-control" value={selectedStrat} onChange={handleStrategyChange} disabled={loading}>
-              {strategies.map(s => <option key={s.name} value={s.name}>{s.name}</option>)}
-            </select>
-          )}
-        </div>
-      </div>
-
-      {currentStrat && currentStrat.parameters && currentStrat.parameters.length > 0 && (
-        <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(160px, 1fr))', gap: 'var(--space-3)' }}>
-          {currentStrat.parameters.map(p => {
-            const val = paramValues[p.name] !== undefined ? paramValues[p.name] : p.default
-            if (p.type === 'bool') {
-              return (
-                <div key={p.name} className="form-group">
-                  <label className="form-label">{p.name}</label>
-                  <div className="toggle-group">
-                    <button type="button" className={`toggle-option${val === true || val === 'true' ? ' active' : ''}`}
-                      onClick={() => !loading && setParamValues(prev => ({ ...prev, [p.name]: true }))}>On</button>
-                    <button type="button" className={`toggle-option${val === false || val === 'false' ? ' active' : ''}`}
-                      onClick={() => !loading && setParamValues(prev => ({ ...prev, [p.name]: false }))}>Off</button>
-                  </div>
-                </div>
-              )
-            }
-            if (p.type === 'str') {
-              return (
-                <div key={p.name} className="form-group">
-                  <label className="form-label">{p.name}</label>
-                  <select className="form-control" value={val}
-                    onChange={e => setParamValues(prev => ({ ...prev, [p.name]: e.target.value }))} disabled={loading}>
-                    {['open_next', 'close_current'].map(o => <option key={o} value={o}>{o}</option>)}
-                  </select>
-                </div>
-              )
-            }
-            return (
-              <div key={p.name} className="form-group">
-                <label className="form-label">{p.name}</label>
-                <input type="number" className="form-control" value={val}
-                  min={p.min ?? undefined} max={p.max ?? undefined}
-                  step={p.type === 'float' ? 0.001 : 1}
-                  onChange={e => {
-                    const parsed = p.type === 'float' ? parseFloat(e.target.value) : parseInt(e.target.value, 10)
-                    setParamValues(prev => ({ ...prev, [p.name]: isNaN(parsed) ? e.target.value : parsed }))
-                  }} disabled={loading} />
-              </div>
-            )
-          })}
-        </div>
-      )}
-
-      <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 'var(--space-4)' }}>
-        <div>
-          <div className="section-title">Risk Parameters</div>
-          <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 'var(--space-3)' }}>
-            <div className="form-group">
-              <label className="form-label">Cost (bps)</label>
-              <input type="number" className="form-control" value={costBps} min={0} step={1}
-                onChange={e => setCostBps(parseFloat(e.target.value) || 0)} disabled={loading} />
-            </div>
-            <div className="form-group">
-              <label className="form-label" title="Maintenance margin used for the liquidation-price formula. Binance baseline ≈ 0.005 (0.5%) for low notional.">Maint. margin %</label>
-              <input type="number" className="form-control" value={maintenanceMarginPct} min={0} step={0.001}
-                onChange={e => setMaintenanceMarginPct(parseFloat(e.target.value) || 0)} disabled={loading} />
-            </div>
-          </div>
-        </div>
-        <div>
-          <div className="section-title">Capital</div>
-          <CapitalConfig
-            portfolio={portfolio} setPortfolio={setPortfolio}
-            leverage={leverage} setLeverage={setLeverage}
-            investedAmount={investedAmount} setInvestedAmount={setInvestedAmount}
-            mode={capitalMode} setMode={setCapitalMode}
-            disabled={loading}
-          />
-        </div>
-      </div>
-
-      {error && (
-        <div style={{
-          padding: '8px 12px', background: 'rgba(239,68,68,0.1)',
-          border: '1px solid rgba(239,68,68,0.25)', borderRadius: 'var(--radius-sm)',
-          color: 'var(--color-danger)', fontSize: '0.83rem',
-        }}>{error}</div>
-      )}
-
-      <div>
-        <button className="btn btn-primary" onClick={handleCreate} disabled={loading || !selectedStrat}>
-          {loading ? 'Creating…' : 'Create Signal Config'}
-        </button>
-      </div>
-    </div>
-  )
-}
-
-/* ---- Configs list ---- */
-function ConfigsList({ configs, onToggle, onToggleTelegram, onDelete, onResetEquity }) {
-  if (!configs || configs.length === 0) {
-    return <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem', padding: 'var(--space-3)' }}>No signal configs yet.</div>
-  }
-
-  return (
-    <div style={{ overflowX: 'auto' }}>
-      <table className="trade-table">
-        <thead>
-          <tr>
-            <th className="ta-right">ID</th>
-            <th>Pair</th>
-            <th>Interval</th>
-            <th>Strategy</th>
-            <th className="ta-right" title="Capital inicial al crear la config (inmutable salvo edición manual)">Inicial</th>
-            <th className="ta-right" title="Capital actual: arranca = inicial y evoluciona con el PnL neto de cada sim-trade cerrado. Se usa para dimensionar nuevos trades.">Actual</th>
-            <th className="ta-right">Δ%</th>
-            <th className="ta-center">Active</th>
-            <th className="ta-center" title="Enviar alertas a Telegram para esta configuración">Telegram</th>
-            <th className="ta-center">Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {configs.map(c => {
-            const initial = Number(c.initial_portfolio ?? 0)
-            const current = Number(c.current_portfolio ?? initial)
-            const deltaPct = initial > 0 ? ((current - initial) / initial) * 100 : 0
-            const deltaColor = deltaPct > 0 ? 'var(--color-success)' : deltaPct < 0 ? 'var(--color-danger)' : 'var(--text-muted)'
-            const blown = c.status === 'blown'
-            return (
-              <tr key={c.id} style={blown ? { opacity: 0.7 } : undefined}>
-                <td className="ta-right num-col" style={{ fontFamily: 'var(--font-mono)', fontSize: '0.8rem' }}>
-                  {c.id}
-                  {blown && (
-                    <span
-                      title={`Blown at ${c.blown_at ?? ''}`}
-                      aria-label={`Account blown at ${c.blown_at ?? 'unknown date'}`}
-                      style={{
-                        marginLeft: 6, padding: '1px 6px', borderRadius: 'var(--radius-sm)',
-                        background: 'rgba(239,68,68,0.2)', color: 'var(--color-danger)',
-                        fontSize: '0.7rem', fontWeight: 700, letterSpacing: '0.03em',
-                      }}>
-                      <span aria-hidden="true">⚠ </span>BLOWN
-                    </span>
-                  )}
-                </td>
-                <td>{c.symbol}</td>
-                <td>{c.interval}</td>
-                <td>{c.strategy}</td>
-                <td className="ta-right num-col">{fmtMoney(initial)}</td>
-                <td className="ta-right num-col" style={{ color: deltaColor, fontWeight: 600 }}>{fmtMoney(current)}</td>
-                <td className="ta-right num-col" style={{ color: deltaColor }}>
-                  {deltaPct === 0 ? '—' : (deltaPct > 0 ? '+' : '') + deltaPct.toFixed(2) + '%'}
-                </td>
-                <td className="ta-center">
-                  <ToggleSwitch
-                    checked={c.active}
-                    onChange={(val) => onToggle(c.id, val)}
-                    disabled={blown}
-                    ariaLabel={`Activar config ${c.id} (${c.symbol} ${c.strategy})`}
-                  />
-                </td>
-                <td className="ta-center">
-                  <ToggleSwitch
-                    checked={!!c.telegram_enabled}
-                    onChange={(val) => onToggleTelegram(c.id, val)}
-                    ariaLabel={`Notificaciones Telegram para config ${c.id} (${c.symbol} ${c.strategy})`}
-                  />
-                </td>
-                <td className="ta-center" style={{ display: 'flex', gap: 4, justifyContent: 'center' }}>
-                  {blown && (
-                    <button className="btn btn-sm btn-secondary"
-                      title="Restore current_portfolio = initial_portfolio and reactivate the config"
-                      onClick={() => { if (confirm('Reset equity to initial and reactivate this config?')) onResetEquity(c.id) }}>
-                      Reset equity
-                    </button>
-                  )}
-                  <button className="btn btn-sm btn-secondary" style={{ color: 'var(--color-danger)' }}
-                    onClick={() => { if (confirm('Delete this config? Open trades will be closed.')) onDelete(c.id) }}>
-                    Delete
-                  </button>
-                </td>
-              </tr>
-            )
-          })}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-/* ---- Signals table ---- */
-function SignalsList({ signals }) {
-  if (!signals || signals.length === 0) {
-    return <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem', padding: 'var(--space-3)' }}>No signals generated yet.</div>
-  }
-  return (
-    <div>
-      {/* Legend */}
-      <div style={{
-        padding: 'var(--space-2) var(--space-3)',
-        background: 'var(--bg-elevated)', borderRadius: 'var(--radius-sm)',
-        marginBottom: 'var(--space-3)', fontSize: '0.76rem', color: 'var(--text-muted)',
-        lineHeight: 1.6,
-      }}>
-        💡 <strong style={{ color: 'var(--text-secondary)' }}>Entry (next open)</strong>: precio al que entrar en el exchange &nbsp;·&nbsp;
-        <strong style={{ color: 'var(--text-secondary)' }}>Stop (SL)</strong>: nivel de stop-loss para tu exchange — el sistema cierra aquí también el SimTrade.
-      </div>
-
-      <div style={{ overflowX: 'auto' }}>
-        <table className="trade-table">
-          <thead>
-            <tr>
-              <th className="ta-right">ID</th>
-              <th>Config</th>
-              <th>Symbol</th>
-              <th>Side</th>
-              <th>Signal Candle</th>
-              <th className="ta-right" title="Open price of the next candle — use this as your entry on the exchange">Entry (next open) ↗</th>
-              <th className="ta-right" title="Strategy stop-loss level — set this as your SL on the exchange and the system closes the SimTrade here too">Stop (SL) 🛑</th>
-              <th>Status</th>
-              <th className="ta-center">Sim Trade</th>
-              <th>Created</th>
-            </tr>
-          </thead>
-          <tbody>
-            {signals.map(s => {
-              const hasEntry = s.entry_price != null
-              const entryColor = hasEntry ? 'var(--text-primary)' : 'var(--color-warning)'
-              const simStatusLabel = s.sim_trade_status
-                ? s.sim_trade_status.replace('_', ' ')
-                : null
-              return (
-                <tr key={s.id}>
-                  <td className="ta-right num-col" style={{ fontFamily: 'var(--font-mono)', fontSize: '0.8rem' }}>{s.id}</td>
-                  <td>
-                    <ConfigBadge configId={s.config_id} strategy={s.strategy} params={s.config_params} />
-                  </td>
-                  <td>{s.symbol}</td>
-                  <td style={{ color: s.side === 'long' ? 'var(--color-success)' : 'var(--color-danger)', fontWeight: 700 }}>
-                    {s.side?.toUpperCase()}
-                  </td>
-                  <td className="num-col" style={{ fontFamily: 'var(--font-mono)', fontSize: '0.78rem' }}>{fmtTime(s.trigger_candle_time)}</td>
-                  <td className="ta-right num-col" style={{ color: entryColor, fontWeight: hasEntry ? 600 : 400 }}>
-                    {hasEntry ? fmtNum(s.entry_price, 4) : 'Pending…'}
-                  </td>
-                  <td className="ta-right num-col" style={{ fontFamily: 'var(--font-mono)' }}>{fmtNum(s.stop_price, 4)}</td>
-                  <td><StatusBadge status={s.status} /></td>
-                  <td className="ta-center" style={{ fontSize: '0.78rem' }}>
-                    {s.sim_trade_id
-                      ? <span>#{s.sim_trade_id}{simStatusLabel ? <span style={{ marginLeft: 4, color: 'var(--text-muted)' }}>({simStatusLabel})</span> : null}</span>
-                      : '—'}
-                  </td>
-                  <td style={{ fontSize: '0.78rem', color: 'var(--text-muted)' }}>{fmtIso(s.created_at)}</td>
-                </tr>
-              )
-            })}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  )
-}
-
-/* ---- SimTrades table ---- */
-function SimTradesList({ trades, onClose }) {
-  const [expandedId, setExpandedId] = useState(null)
-  const [movesByTrade, setMovesByTrade] = useState({})
-  const [loadingMoves, setLoadingMoves] = useState(false)
-
-  const toggleExpand = useCallback(async (tradeId) => {
-    if (expandedId === tradeId) {
-      setExpandedId(null)
-      return
-    }
-    setExpandedId(tradeId)
-    if (movesByTrade[tradeId] !== undefined) return
-    setLoadingMoves(true)
-    try {
-      const res = await apiFetch(`/api/sim-trades/${tradeId}/stop-moves`)
-      if (res.ok) {
-        const data = await res.json()
-        setMovesByTrade(prev => ({ ...prev, [tradeId]: data.stop_moves ?? [] }))
-      } else {
-        setMovesByTrade(prev => ({ ...prev, [tradeId]: [] }))
-      }
-    } catch {
-      setMovesByTrade(prev => ({ ...prev, [tradeId]: [] }))
-    } finally {
-      setLoadingMoves(false)
-    }
-  }, [expandedId, movesByTrade])
-
-  if (!trades || trades.length === 0) {
-    return <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem', padding: 'var(--space-3)' }}>No sim trades yet.</div>
-  }
-  return (
-    <div style={{ overflowX: 'auto' }}>
-      <table className="trade-table">
-        <thead>
-          <tr>
-            <th style={{ width: '2rem' }}></th>
-            <th className="ta-right">ID</th>
-            <th>Config</th>
-            <th>Symbol</th>
-            <th>Side</th>
-            <th className="ta-right" title="Capital usado al abrir este trade (snapshot del current_portfolio en ese momento)">Cap@entry</th>
-            <th className="ta-right">Entry</th>
-            <th className="ta-right">Stop</th>
-            <th className="ta-right" title="Precio de liquidación. NULL en trades sin apalancamiento. Se cierra automáticamente aquí en vez de en el stop si el precio cruza primero.">Liq</th>
-            <th className="ta-right">Exit</th>
-            <th>Reason</th>
-            <th className="ta-right">PnL</th>
-            <th className="ta-right">PnL %</th>
-            <th>Status</th>
-            <th className="ta-center">Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {trades.map(t => {
-            const pnlColor = t.pnl > 0 ? 'var(--color-success)' : t.pnl < 0 ? 'var(--color-danger)' : 'var(--text-secondary)'
-            const isExpanded = expandedId === t.id
-            const moves = movesByTrade[t.id]
-            return (
-              <React.Fragment key={t.id}>
-                <tr>
-                  <td className="ta-center">
-                    <button
-                      type="button"
-                      onClick={() => toggleExpand(t.id)}
-                      title="Show stop-move history"
-                      style={{
-                        background: 'transparent', border: 'none', cursor: 'pointer',
-                        color: 'var(--text-muted)', padding: '0.25rem',
-                      }}
-                    >{isExpanded ? '▾' : '▸'}</button>
-                  </td>
-                  <td className="ta-right num-col" style={{ fontFamily: 'var(--font-mono)', fontSize: '0.8rem' }}>{t.id}</td>
-                  <td>
-                    <ConfigBadge configId={t.config_id} strategy={t.config_strategy} params={t.config_params} />
-                  </td>
-                  <td>{t.symbol}</td>
-                  <td style={{ color: t.side === 'long' ? 'var(--color-success)' : 'var(--color-danger)', fontWeight: 600 }}>
-                    {t.side?.toUpperCase()}
-                  </td>
-                  <td className="ta-right num-col" style={{ color: 'var(--text-muted)', fontSize: '0.8rem' }}>{t.portfolio != null ? fmtMoney(t.portfolio) : '—'}</td>
-                  <td className="ta-right num-col">{t.entry_price ? fmtNum(t.entry_price, 4) : '—'}</td>
-                  <td className="ta-right num-col">{fmtNum(t.stop_base, 4)}</td>
-                  <td className="ta-right num-col" style={{ color: t.liquidation_price != null ? 'var(--color-warning)' : 'var(--text-muted)', fontSize: '0.85rem' }}>
-                    {t.liquidation_price != null ? fmtNum(t.liquidation_price, 4) : '—'}
-                  </td>
-                  <td className="ta-right num-col">{t.exit_price ? fmtNum(t.exit_price, 4) : '—'}</td>
-                  <td>{t.exit_reason || '—'}</td>
-                  <td className="ta-right num-col" style={{ color: pnlColor, fontWeight: 600 }}>
-                    {t.pnl != null ? (t.pnl > 0 ? '+' : '') + fmtMoney(t.pnl) : '—'}
-                  </td>
-                  <td className="ta-right num-col" style={{ color: pnlColor }}>
-                    {t.pnl_pct != null ? (t.pnl_pct > 0 ? '+' : '') + fmtNum(t.pnl_pct * 100, 2) + '%' : '—'}
-                  </td>
-                  <td><StatusBadge status={t.status} /></td>
-                  <td className="ta-center">
-                    {t.status === 'open' && (
-                      <button className="btn btn-sm btn-secondary" onClick={() => onClose(t.id)}>Close</button>
-                    )}
-                  </td>
-                </tr>
-                {isExpanded && (
-                  <tr>
-                    <td colSpan={15} style={{ background: 'var(--surface-1)', padding: 'var(--space-3)' }}>
-                      <SimTradeReview trade={t} moves={moves} loadingMoves={loadingMoves && moves === undefined} />
-                    </td>
-                  </tr>
-                )}
-              </React.Fragment>
-            )
-          })}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-function SimTradeReview({ trade, moves, loadingMoves }) {
-  // Capture "now" once at mount so the render function stays pure for the
-  // React Compiler. Open trades (no exit_time) chart up to this snapshot.
-  const [nowMs] = useState(() => Date.now())
-
-  const { startMs, endMs } = useMemo(() => {
-    const stepBy = {
-      '1h': 60 * 60 * 1000,
-      '2h': 2 * 60 * 60 * 1000,
-      '4h': 4 * 60 * 60 * 1000,
-      '6h': 6 * 60 * 60 * 1000,
-      '8h': 8 * 60 * 60 * 1000,
-      '12h': 12 * 60 * 60 * 1000,
-      '1d': 24 * 60 * 60 * 1000,
-      '3d': 3 * 24 * 60 * 60 * 1000,
-      '1w': 7 * 24 * 60 * 60 * 1000,
-      '1M': 30 * 24 * 60 * 60 * 1000,
-    }
-    const STEP_MS = stepBy[trade.interval] || 24 * 60 * 60 * 1000
-    const entryMs = Number(trade.entry_time) || nowMs
-    const exitMs = Number(trade.exit_time) || nowMs
-    return {
-      startMs: Math.max(0, entryMs - 50 * STEP_MS),
-      endMs: exitMs + 10 * STEP_MS,
-    }
-  }, [trade.interval, trade.entry_time, trade.exit_time, nowMs])
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--space-3)' }}>
-      {trade.entry_time && (
-        <TradeReviewChart
-          symbol={trade.symbol}
-          interval={trade.interval}
-          startMs={startMs}
-          endMs={endMs}
-          trades={[{
-            id: trade.id,
-            side: trade.side,
-            entry_time: trade.entry_time,
-            exit_time: trade.exit_time ?? endMs,
-            entry_price: trade.entry_price,
-            exit_price: trade.exit_price,
-            pnl: trade.pnl,
-            exit_reason: trade.exit_reason,
-            stop_base: trade.stop_base,
-          }]}
-        />
-      )}
-      <StopMovesDetail moves={moves} loading={loadingMoves} />
-    </div>
-  )
-}
-
-function StopMovesDetail({ moves, loading }) {
-  if (loading) return <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem' }}>Loading stop moves…</div>
-  if (!moves || moves.length === 0) {
-    return <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem' }}>No stop moves for this trade.</div>
-  }
-  return (
-    <div>
-      <div style={{ fontWeight: 600, marginBottom: 'var(--space-2)' }}>Stop moves ({moves.length})</div>
-      <table className="trade-table" style={{ fontSize: '0.8rem' }}>
-        <thead>
-          <tr>
-            <th className="ta-right">#</th>
-            <th className="ta-right">Stop anterior</th>
-            <th className="ta-right">Stop nuevo</th>
-            <th className="ta-right">Candle time</th>
-            <th>Created at</th>
-          </tr>
-        </thead>
-        <tbody>
-          {moves.map((m, i) => (
-            <tr key={m.id}>
-              <td className="ta-right num-col">{i + 1}</td>
-              <td className="ta-right num-col">{fmtNum(m.prev_stop_base, 4)}</td>
-              <td className="ta-right num-col">{fmtNum(m.new_stop_base, 4)}</td>
-              <td className="ta-right num-col">{new Date(m.candle_time).toISOString().replace('T', ' ').slice(0, 19)}</td>
-              <td>{m.created_at}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-/* ---- Real Trades section ---- */
-function RealTradesSection() {
-  const [realTrades, setRealTrades] = useState([])
-  const [showForm, setShowForm] = useState(false)
-  const [form, setForm] = useState({
-    sim_trade_id: '', symbol: '', side: 'long', entry_price: '', entry_time: '', quantity: '', fees: '0', notes: '',
-  })
-  const [loading, setLoading] = useState(false)
-  const [closingId, setClosingId] = useState(null)
-  const [closeForm, setCloseForm] = useState({ exit_price: '', pnl: '', exit_time: '', notes: '' })
-  const [closeError, setCloseError] = useState(null)
-  const [justClosedId, setJustClosedId] = useState(null)
-
-  const fetchRealTrades = useCallback(async () => {
-    try {
-      const res = await apiFetch('/api/real-trades?limit=100')
-      if (res.ok) {
-        const data = await res.json()
-        setRealTrades(data.real_trades ?? [])
-      }
-    } catch { /* ignore */ }
-  }, [])
-
-  // eslint-disable-next-line react-hooks/set-state-in-effect
-  useEffect(() => { fetchRealTrades() }, [fetchRealTrades])
-
-  const handleCreate = async () => {
-    setLoading(true)
-    try {
-      const body = {
-        symbol: form.symbol || 'BTCUSDT',
-        side: form.side,
-        entry_price: parseFloat(form.entry_price),
-        entry_time: form.entry_time || new Date().toISOString(),
-        quantity: parseFloat(form.quantity),
-        fees: parseFloat(form.fees) || 0,
-        notes: form.notes || null,
-        sim_trade_id: form.sim_trade_id ? parseInt(form.sim_trade_id) : null,
-      }
-      const res = await apiFetch('/api/real-trades', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      })
-      if (res.ok) {
-        setShowForm(false)
-        setForm({ sim_trade_id: '', symbol: '', side: 'long', entry_price: '', entry_time: '', quantity: '', fees: '0', notes: '' })
-        fetchRealTrades()
-      }
-    } catch { /* ignore */ }
-    setLoading(false)
-  }
-
-  const FEE_WARN_PCT = 0.05
-
-  const computeCloseSummary = (trade) => {
-    const exitPrice = parseFloat(closeForm.exit_price)
-    const netPnl = parseFloat(closeForm.pnl)
-    const qty = parseFloat(trade.quantity)
-    const entry = parseFloat(trade.entry_price)
-    if ([exitPrice, netPnl, qty, entry].some(v => !Number.isFinite(v))) return null
-    const invested = qty * entry
-    const gross = trade.side === 'long' ? qty * (exitPrice - entry) : qty * (entry - exitPrice)
-    const fee = gross - netPnl
-    const feePct = invested > 0 ? Math.abs(fee) / invested : 0
-    const warn = fee < 0 || feePct > FEE_WARN_PCT
-    return { invested, gross, netPnl, fee, feePct, warn }
-  }
-
-  const handleClose = async (trade) => {
-    setLoading(true)
-    setCloseError(null)
-    try {
-      const s = computeCloseSummary(trade)
-      const body = {
-        exit_price: parseFloat(closeForm.exit_price),
-        exit_time: closeForm.exit_time || new Date().toISOString(),
-        pnl: parseFloat(closeForm.pnl),
-        fees: s?.fee ?? 0,
-        status: 'closed',
-      }
-      if (closeForm.notes) body.notes = closeForm.notes
-      const res = await apiFetch(`/api/real-trades/${trade.id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      })
-      if (!res.ok) {
-        let detail = `HTTP ${res.status}`
-        try { const j = await res.json(); if (j?.detail) detail = j.detail } catch { /* non-JSON body */ }
-        setCloseError(`Close failed: ${detail}`)
-        return
-      }
-      await fetchRealTrades()
-      setClosingId(null)
-      setCloseForm({ exit_price: '', pnl: '', exit_time: '', notes: '' })
-      setJustClosedId(trade.id)
-      setTimeout(() => setJustClosedId(prev => (prev === trade.id ? null : prev)), 1500)
-    } catch (e) {
-      setCloseError(e?.message || 'Network error closing trade')
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const openCloseForm = (trade) => {
-    setClosingId(trade.id)
-    setCloseForm({ exit_price: '', pnl: '', exit_time: '', notes: '' })
-    setCloseError(null)
-  }
-
-  return (
-    <div>
-      <div style={{ display: 'flex', alignItems: 'center', gap: 'var(--space-3)', marginBottom: 'var(--space-3)' }}>
-        <button className="btn btn-sm btn-primary" onClick={() => setShowForm(!showForm)}>
-          {showForm ? 'Cancel' : '+ Register Real Trade'}
-        </button>
-        <button className="btn btn-sm btn-secondary" onClick={fetchRealTrades}>Refresh</button>
-      </div>
-
-      {showForm && (
-        <div className="card" style={{ marginBottom: 'var(--space-4)', padding: 'var(--space-4)' }}>
-          <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(150px, 1fr))', gap: 'var(--space-3)' }}>
-            <div className="form-group">
-              <label className="form-label">Link to SimTrade ID</label>
-              <input type="number" className="form-control" value={form.sim_trade_id}
-                onChange={e => setForm(prev => ({ ...prev, sim_trade_id: e.target.value }))} placeholder="Optional" />
-            </div>
-            <div className="form-group">
-              <label className="form-label">Symbol</label>
-              <select className="form-control" value={form.symbol}
-                onChange={e => setForm(prev => ({ ...prev, symbol: e.target.value }))}>
-                {PAIRS.map(p => <option key={p} value={p}>{p}</option>)}
-              </select>
-            </div>
-            <div className="form-group">
-              <label className="form-label">Side</label>
-              <select className="form-control" value={form.side}
-                onChange={e => setForm(prev => ({ ...prev, side: e.target.value }))}>
-                <option value="long">Long</option>
-                <option value="short">Short</option>
-              </select>
-            </div>
-            <div className="form-group">
-              <label className="form-label">Entry Price</label>
-              <input type="number" className="form-control" value={form.entry_price} step="0.01"
-                onChange={e => setForm(prev => ({ ...prev, entry_price: e.target.value }))} />
-            </div>
-            <div className="form-group">
-              <label className="form-label">Quantity</label>
-              <input type="number" className="form-control" value={form.quantity} step="0.0001"
-                onChange={e => setForm(prev => ({ ...prev, quantity: e.target.value }))} />
-            </div>
-            <div className="form-group">
-              <label className="form-label">Fees</label>
-              <input type="number" className="form-control" value={form.fees} step="0.01"
-                onChange={e => setForm(prev => ({ ...prev, fees: e.target.value }))} />
-            </div>
-          </div>
-          <div className="form-group" style={{ marginTop: 'var(--space-3)' }}>
-            <label className="form-label">Notes</label>
-            <input type="text" className="form-control" value={form.notes}
-              onChange={e => setForm(prev => ({ ...prev, notes: e.target.value }))} placeholder="Optional notes" />
-          </div>
-          <button className="btn btn-primary btn-sm" style={{ marginTop: 'var(--space-3)' }}
-            onClick={handleCreate} disabled={loading || !form.entry_price || !form.quantity}>
-            Save Real Trade
-          </button>
-        </div>
-      )}
-
-      {realTrades.length === 0 ? (
-        <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem' }}>No real trades registered.</div>
-      ) : (
-        <div style={{ overflowX: 'auto' }}>
-          <table className="trade-table">
-            <thead>
-              <tr>
-                <th className="ta-right">ID</th>
-                <th className="ta-right">Sim #</th>
-                <th>Symbol</th>
-                <th>Side</th>
-                <th className="ta-right">Entry</th>
-                <th className="ta-right">Exit</th>
-                <th className="ta-right">PnL</th>
-                <th className="ta-right">Fees</th>
-                <th>Status</th>
-                <th>Notes</th>
-                <th className="ta-center">Actions</th>
-              </tr>
-            </thead>
-            <tbody>
-              {realTrades.map(t => {
-                const pnlColor = t.pnl > 0 ? 'var(--color-success)' : t.pnl < 0 ? 'var(--color-danger)' : 'var(--text-secondary)'
-                return (
-                  <React.Fragment key={t.id}>
-                  <tr style={{
-                    transition: 'background 300ms ease',
-                    background: t.id === justClosedId ? 'rgba(34,197,94,0.18)' : undefined,
-                  }}>
-                    <td className="ta-right num-col" style={{ fontFamily: 'var(--font-mono)', fontSize: '0.8rem' }}>{t.id}</td>
-                    <td className="ta-right num-col">{t.sim_trade_id || '—'}</td>
-                    <td>{t.symbol}</td>
-                    <td style={{ color: t.side === 'long' ? 'var(--color-success)' : 'var(--color-danger)', fontWeight: 600 }}>
-                      {t.side?.toUpperCase()}
-                    </td>
-                    <td className="ta-right num-col">{fmtNum(t.entry_price, 4)}</td>
-                    <td className="ta-right num-col">{t.exit_price ? fmtNum(t.exit_price, 4) : '—'}</td>
-                    <td className="ta-right num-col" style={{ color: pnlColor, fontWeight: 600 }}>{t.pnl != null ? fmtMoney(t.pnl) : '—'}</td>
-                    <td className="ta-right num-col">{fmtNum(t.fees, 2)}</td>
-                    <td><StatusBadge status={t.status} /></td>
-                    <td style={{ fontSize: '0.8rem', maxWidth: 150, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-                      {t.notes || '—'}
-                    </td>
-                    <td className="ta-center">
-                      {t.status === 'open' && closingId !== t.id && (
-                        <button className="btn btn-sm" onClick={() => openCloseForm(t)}
-                          style={{ background: 'var(--color-warning)', color: '#000', fontSize: '0.75rem', padding: '2px 8px' }}>
-                          Close
-                        </button>
-                      )}
-                      {closingId === t.id && (
-                        <button className="btn btn-sm btn-secondary" onClick={() => setClosingId(null)}
-                          style={{ fontSize: '0.75rem', padding: '2px 8px' }}>
-                          Cancel
-                        </button>
-                      )}
-                    </td>
-                  </tr>
-                  {closingId === t.id && (() => {
-                    const s = computeCloseSummary(t)
-                    return (
-                      <tr key={`close-${t.id}`} style={{ background: 'var(--bg-elevated)' }}>
-                        <td colSpan={11} style={{ padding: 'var(--space-3)' }}>
-                          <div style={{ display: 'flex', gap: 'var(--space-3)', alignItems: 'flex-end', flexWrap: 'wrap' }}>
-                            <div className="form-group" style={{ marginBottom: 0 }}>
-                              <label className="form-label" style={{ fontSize: '0.75rem' }}>Exit Price *</label>
-                              <input type="number" className="form-control" step="0.01" style={{ width: 130 }}
-                                value={closeForm.exit_price}
-                                onChange={e => setCloseForm(prev => ({ ...prev, exit_price: e.target.value }))} />
-                            </div>
-                            <div className="form-group" style={{ marginBottom: 0 }}>
-                              <label className="form-label" style={{ fontSize: '0.75rem' }}>Net PnL *</label>
-                              <input type="number" className="form-control" step="0.01" style={{ width: 130 }}
-                                value={closeForm.pnl}
-                                onChange={e => setCloseForm(prev => ({ ...prev, pnl: e.target.value }))} />
-                            </div>
-                            <div className="form-group" style={{ marginBottom: 0 }}>
-                              <label className="form-label" style={{ fontSize: '0.75rem' }}>Exit Time</label>
-                              <input type="datetime-local" className="form-control" style={{ width: 190 }}
-                                value={closeForm.exit_time}
-                                onChange={e => setCloseForm(prev => ({ ...prev, exit_time: e.target.value }))} />
-                            </div>
-                            <div className="form-group" style={{ marginBottom: 0 }}>
-                              <label className="form-label" style={{ fontSize: '0.75rem' }}>Notes</label>
-                              <input type="text" className="form-control" style={{ width: 160 }}
-                                value={closeForm.notes} placeholder="Optional"
-                                onChange={e => setCloseForm(prev => ({ ...prev, notes: e.target.value }))} />
-                            </div>
-                            <button className="btn btn-sm btn-primary" onClick={() => handleClose(t)}
-                              disabled={loading || !closeForm.exit_price || !closeForm.pnl}
-                              style={{ fontSize: '0.75rem' }}>
-                              {loading ? 'Closing…' : 'Confirm Close'}
-                            </button>
-                          </div>
-                          {s && (
-                            <div style={{ marginTop: 'var(--space-2)', display: 'flex', gap: 'var(--space-4)',
-                                          flexWrap: 'wrap', fontSize: '0.8rem',
-                                          fontFamily: 'var(--font-mono)', color: 'var(--text-secondary)' }}>
-                              <span>Invertido: {fmtMoney(s.invested)}</span>
-                              <span>Gross: {fmtMoney(s.gross)}</span>
-                              <span style={{ color: s.warn ? 'var(--color-warning)' : undefined }}>
-                                Fee: {fmtMoney(s.fee)} ({(s.feePct * 100).toFixed(2)}%)
-                              </span>
-                            </div>
-                          )}
-                          {s?.warn && (
-                            <div style={{ marginTop: 'var(--space-2)', padding: '6px 10px',
-                                          background: 'rgba(234,179,8,0.1)',
-                                          border: '1px solid rgba(234,179,8,0.25)',
-                                          borderRadius: 'var(--radius-sm)',
-                                          color: 'var(--color-warning)', fontSize: '0.8rem' }}>
-                              {s.fee < 0
-                                ? 'Fee negativo: el Net PnL que has puesto es mayor que el Gross teórico — revisa el signo o la escala.'
-                                : `Fee = ${(s.feePct * 100).toFixed(2)}% del invertido: inusualmente alto. Revisa el Net PnL.`}
-                            </div>
-                          )}
-                          {closeError && (
-                            <div style={{ marginTop: 'var(--space-2)', padding: '6px 10px',
-                                          background: 'rgba(239,68,68,0.1)',
-                                          border: '1px solid rgba(239,68,68,0.25)',
-                                          borderRadius: 'var(--radius-sm)',
-                                          color: 'var(--color-danger)', fontSize: '0.8rem' }}>
-                              {closeError}
-                            </div>
-                          )}
-                        </td>
-                      </tr>
-                    )
-                  })()}
-                  </React.Fragment>
-                )
-              })}
-            </tbody>
-          </table>
-        </div>
-      )}
-    </div>
-  )
-}
-
-/* ---- Comparison view ---- */
-function ComparisonView() {
-  const [simId, setSimId] = useState('')
-  const [data, setData] = useState(null)
-  const [loading, setLoading] = useState(false)
-
-  const handleFetch = async () => {
-    if (!simId) return
-    setLoading(true)
-    try {
-      const res = await apiFetch(`/api/comparison/${simId}`)
-      if (res.ok) {
-        setData(await res.json())
-      }
-    } catch { /* ignore */ }
-    setLoading(false)
-  }
-
-  return (
-    <div>
-      <div style={{ display: 'flex', gap: 'var(--space-3)', alignItems: 'flex-end', marginBottom: 'var(--space-4)' }}>
-        <div className="form-group" style={{ marginBottom: 0 }}>
-          <label className="form-label">SimTrade ID</label>
-          <input type="number" className="form-control" value={simId} onChange={e => setSimId(e.target.value)}
-            style={{ width: 120 }} />
-        </div>
-        <button className="btn btn-primary btn-sm" onClick={handleFetch} disabled={loading || !simId}>
-          Compare
-        </button>
-      </div>
-
-      {data && (
-        <div>
-          <div className="section-title">SimTrade #{data.sim_trade?.id}</div>
-          <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(140px, 1fr))', gap: 'var(--space-2)', marginBottom: 'var(--space-4)' }}>
-            <div className="metric-card">
-              <span className="metric-label">Side</span>
-              <span className="metric-value">{data.sim_trade?.side?.toUpperCase()}</span>
-            </div>
-            <div className="metric-card">
-              <span className="metric-label">Entry</span>
-              <span className="metric-value">{fmtNum(data.sim_trade?.entry_price, 4)}</span>
-            </div>
-            <div className="metric-card">
-              <span className="metric-label">Exit</span>
-              <span className="metric-value">{data.sim_trade?.exit_price ? fmtNum(data.sim_trade.exit_price, 4) : '—'}</span>
-            </div>
-            <div className="metric-card">
-              <span className="metric-label">PnL</span>
-              <span className={`metric-value ${(data.sim_trade?.pnl ?? 0) >= 0 ? 'positive' : 'negative'}`}>
-                {data.sim_trade?.pnl != null ? fmtMoney(data.sim_trade.pnl) : '—'}
-              </span>
-            </div>
-            <div className="metric-card">
-              <span className="metric-label">Reason</span>
-              <span className="metric-value">{data.sim_trade?.exit_reason || '—'}</span>
-            </div>
-          </div>
-
-          {data.comparisons?.length > 0 ? data.comparisons.map((comp, i) => (
-            <div key={i} style={{ marginBottom: 'var(--space-4)' }}>
-              <div className="section-title">vs Real Trade #{comp.real_trade?.id}</div>
-              <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(160px, 1fr))', gap: 'var(--space-2)' }}>
-                <div className="metric-card">
-                  <span className="metric-label">Entry Slippage</span>
-                  <span className="metric-value">{comp.entry_slippage != null ? fmtNum(comp.entry_slippage, 6) : '—'}</span>
-                </div>
-                <div className="metric-card">
-                  <span className="metric-label">Exit Slippage</span>
-                  <span className="metric-value">{comp.exit_slippage != null ? fmtNum(comp.exit_slippage, 6) : '—'}</span>
-                </div>
-                <div className="metric-card">
-                  <span className="metric-label">PnL Difference</span>
-                  <span className={`metric-value ${(comp.pnl_diff ?? 0) >= 0 ? 'positive' : 'negative'}`}>
-                    {comp.pnl_diff != null ? fmtMoney(comp.pnl_diff) : '—'}
-                  </span>
-                </div>
-                <div className="metric-card">
-                  <span className="metric-label">Real PnL</span>
-                  <span className="metric-value">{comp.real_trade?.pnl != null ? fmtMoney(comp.real_trade.pnl) : '—'}</span>
-                </div>
-              </div>
-            </div>
-          )) : (
-            <div style={{ color: 'var(--text-muted)', fontSize: '0.85rem' }}>No real trades linked to this SimTrade.</div>
-          )}
-        </div>
-      )}
-    </div>
-  )
-}
-
-/* ==================================================================
-   Main SignalsPanel
-   ================================================================== */
 export default function SignalsPanel() {
   const [tab, setTab] = useState('configs')
   const [strategies, setStrategies] = useState([])
@@ -1237,7 +121,6 @@ export default function SignalsPanel() {
   return (
     <div className="panel-section">
 
-      {/* Status bar */}
       {status && (
         <div style={{
           display: 'flex', gap: 'var(--space-4)', padding: 'var(--space-3) var(--space-4)',
@@ -1254,7 +137,6 @@ export default function SignalsPanel() {
         </div>
       )}
 
-      {/* Sub-tabs */}
       <div style={{ display: 'flex', gap: 'var(--space-2)', marginBottom: 'var(--space-4)', flexWrap: 'wrap' }}>
         {TABS.map(t => (
           <button
@@ -1265,7 +147,6 @@ export default function SignalsPanel() {
         ))}
       </div>
 
-      {/* Tab content */}
       {tab === 'configs' && (
         <div className="card">
           <div className="card-header">
@@ -1273,7 +154,7 @@ export default function SignalsPanel() {
           </div>
           <div className="card-body">
             <div className="section-title">Active Configurations</div>
-            <ConfigsList configs={configs} onToggle={handleToggle} onToggleTelegram={handleToggleTelegram} onDelete={handleDelete} onResetEquity={handleResetEquity} onRefresh={fetchConfigs} />
+            <ConfigsList configs={configs} onToggle={handleToggle} onToggleTelegram={handleToggleTelegram} onDelete={handleDelete} onResetEquity={handleResetEquity} />
             <hr className="divider" />
             <div className="section-title">New Configuration</div>
             <ConfigForm strategies={strategies} onCreated={refreshAll} />
@@ -1309,7 +190,7 @@ export default function SignalsPanel() {
             <span className="card-title">Real Trades</span>
           </div>
           <div className="card-body">
-            <RealTradesSection simTrades={simTrades} />
+            <RealTradesSection />
           </div>
         </div>
       )}
