+++ frontend/src/components/signals/ToggleSwitch.jsx
@@ -1,0 +1,25 @@
+export function ToggleSwitch({ checked, onChange, disabled, ariaLabel }) {
+  return (
+    <button
+      type="button"
+      role="switch"
+      aria-checked={!!checked}
+      aria-label={ariaLabel}
+      aria-disabled={disabled || undefined}
+      disabled={disabled}
+      onClick={() => !disabled && onChange(!checked)}
+      style={{
+        width: 44, height: 24, borderRadius: 12, border: 'none', cursor: disabled ? 'default' : 'pointer',
+        background: checked ? 'var(--color-success)' : 'var(--border-strong)',
+        position: 'relative', transition: 'background var(--transition-fast)',
+        opacity: disabled ? 0.5 : 1,
+      }}
+    >
+      <span style={{
+        position: 'absolute', top: 2, left: checked ? 22 : 2,
+        width: 20, height: 20, borderRadius: '50%', background: 'var(--text-primary)',
+        transition: 'left var(--transition-fast)',
+      }} />
+    </button>
+  )
+}
